web_app/requirements.txt
flask
flask-login
flask-pymongo
werkzeug
Pillow
pytest
pytest-cov